infra: Split container.hpp into container/index_map.hpp, container/reserved_heap.hpp and container/reserved_vec.hpp
