+++ CMakeLists.txt
@@ -15,17 +15,19 @@ endif()
 
 
 set(INFRA_SOURCES
+	infra/container/index_map.hpp
+	infra/container/reserved_heap.hpp
+	infra/container/reserved_vec.hpp
+	infra/minos.hpp
+	infra/minos_win32.cpp
+	infra/minos_linux.cpp
 	infra/alloc_pool.hpp
 	infra/alloc_pool.cpp
 	infra/common.hpp
 	infra/common.cpp
 	infra/hash.hpp
 	infra/inplace_sort.hpp
-	infra/minos.hpp
-	infra/minos_win32.cpp
-	infra/minos_linux.cpp
 	infra/range.hpp
-	infra/container.hpp
 	infra/threading.hpp
 )
