+++ core/global_value_pool.cpp
@@ -1,6 +1,6 @@
 #include "core.hpp"
 
-#include "../infra/container.hpp"
+#include "../infra/container/reserved_vec.hpp"
 
 struct alignas(8) ValueInfo
 {
