+++ core/source_reader.cpp
@@ -1,7 +1,7 @@
 #include "core.hpp"
 
 #include "../infra/common.hpp"
-#include "../infra/container.hpp"
+#include "../infra/container/index_map.hpp"
 #include "../infra/threading.hpp"
 #include "../infra/minos.hpp"
 #include "../infra/hash.hpp"
