+++ core/type_pool.cpp
@@ -1,7 +1,8 @@
 #include "core.hpp"
 
 #include "../infra/hash.hpp"
-#include "../infra/container.hpp"
+#include "../infra/container/index_map.hpp"
+#include "../infra/container/reserved_heap.hpp"
 #include "../infra/inplace_sort.hpp"
 
 #include <cstdlib>
